# Accept double-quotes, comments in selectors