@Krinkle 2020-08-23
### Benchmark

Before:
```
nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3229 op/s in 0.3s
Bench ooui: 645 op/s in 1.5s

nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3176 op/s in 0.3s
Bench ooui: 636 op/s in 1.6s

nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3198 op/s in 0.3s
Bench ooui: 637 op/s in 1.6s
```

After:
```
nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3140 op/s in 0.3s
Bench ooui: 630 op/s in 1.6s

nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3197 op/s in 0.3s
Bench ooui: 644 op/s in 1.6s

nobody@0c6fcfefbc0c:/cssjanus$ node test/bench.js 
Bench mediawiki: 3243 op/s in 0.3s
Bench ooui: 644 op/s in 1.6s
```